+++ mountpoint-s3/tests/fuse_tests/mod.rs
@@ -13,3 +13,4 @@ mod semantics_doc_test;
 mod setattr_test;
 mod unlink_test;
 mod write_test;
+mod statfs_test;